some fixes
-set organization_id in user/login API for users
-fix user/update doesn't check User-Token
-fix user/create doesn't send activation email after creating EmployeeUser

+++ src/main/java/com/nasnav/service/CommonUserServiceInterface.java
@@ -63,5 +63,5 @@ public interface CommonUserServiceInterface {
      * @param authToken token generated on log-in
      * @return true if user is authenticated (token is valid)
      */
-    boolean checkAuthToken(long userId, String authToken);
+    boolean checkAuthToken(Integer userId, String authToken);
 }
